@@ -1,62 +1,81 @@
 {
-  "name": "tapis-ui-cra",
+  "name": "@tapis/tapisui",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.25",
-    "@fortawesome/free-regular-svg-icons": "^5.11.2",
-    "@fortawesome/free-solid-svg-icons": "^5.11.2",
-    "@fortawesome/react-fontawesome": "^0.1.8",
-    "@material-ui/core": "^4.12.3",
-    "@reduxjs/toolkit": "^1.5.1",
-    "@tapis/tapis-typescript": "^0.0.18",
-    "axios": "^0.21.4",
-    "babel-plugin-react-css-modules": "^5.2.6",
-    "bootstrap": "^4.6.0",
-    "cross-fetch": "^3.1.4",
+    "@babel/runtime": "^7.25.0",
+    "@codemirror/lang-json": "^6.0.1",
+    "@codemirror/lang-python": "^6.1.6",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@icicle/tapisui-extension": "file:./packages/icicle-tapisui-extension",
+    "@mui/icons-material": "^5.15.20",
+    "@mui/lab": "^5.0.0-alpha.173",
+    "@mui/material": "^5.16.0",
+    "@mui/x-data-grid": "^7.7.0",
+    "@mui/x-tree-view": "^7.16.0",
+    "@reduxjs/toolkit": "^2.2.7",
+    "@tapis/tapis-typescript": "^0.0.48",
+    "@tapis/tapisui-api": "file:./packages/tapisui-api",
+    "@tapis/tapisui-common": "file:./packages/tapisui-common",
+    "@tapis/tapisui-extensions-core": "file:./packages/tapisui-extensions-core",
+    "@tapis/tapisui-hooks": "file:./packages/tapisui-hooks",
+    "@uiw/codemirror-theme-vscode": "^4.22.2",
+    "@uiw/react-codemirror": "^4.19.7",
+    "@uiw/react-textarea-code-editor": "^2.0.6",
+    "@vitejs/plugin-react": "^4.3.0",
+    "@xyflow/react": "^12.0.4",
+    "base-64": "^1.0.0",
+    "bootstrap": "^4.6.2",
+    "date-fns": "^3.6.0",
+    "elkjs": "^0.9.3",
     "formik": "^2.2.9",
     "js-cookie": "^3.0.0",
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
+    "markdown-to-jsx": "^7.4.7",
     "prop-types": "^15.7.2",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-copy-to-clipboard": "^5.0.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-dropzone": "^10.2.1",
-    "react-hot-loader": "^4.13.0",
     "react-query": "^3.19.2",
-    "react-resize-detector": "^6.1.0",
+    "react-redux": "^9.1.2",
+    "react-resize-detector": "^11.0.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
     "react-step-wizard": "^5.3.9",
     "react-table": "^7.0.5",
     "react-transition-group": "^4.4.1",
-    "reactstrap": "^8.9.0",
+    "reactflow": "^11.4.2",
+    "reactstrap": "^8.10.1",
     "rxjs": "^6.6.7",
     "sass": "^1.37.5",
-    "serve": "^14.1.2",
     "streamsaver": "^2.0.5",
-    "typescript": "^4.3.5",
     "uuid": "^7.0.2",
+    "vite": "^5.2.12",
     "web-vitals": "^1.1.2",
     "yup": "^0.32.11"
   },
   "homepage": "https://tapis-project.github.io/tapis-ui",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "TZ=UTC react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite --config ./vite/vite.config.mts",
+    "start-dev": "vite --host --config ./vite/vite.dev.config.mts",
+    "init-project": "chmod +x ./scripts/init-project.sh && ./scripts/init-project.sh",
+    "watcher": "node ./scripts/watcher.js",
+    "dev": "npm run watcher & npm run start",
+    "build": "tsc && NODE_OPTIONS=--max-old-space-size=9000 vite build --config ./vite/vite.config.mts",
+    "build-dev": "tsc && NODE_OPTIONS=--max-old-space-size=9000 vite build --config ./vite/vite.dev.config.mts",
+    "preview": "vite preview --config ./vite/vite.config.mts",
+    "preview-dev": "vite preview --config ./vite/vite.dev.config.mts",
+    "docker": "docker build -t uidev -f Dockerfile.dev .; docker run -p 3000:3000 uidev",
+    "docker-prod": "docker build -t uiprod .; docker run -p 3000:80 uiprod",
+    "test": "chmod +x ./scripts/test-all.sh && ./scripts/test-all.sh",
     "lint": "eslint --max-warnings=0 . && prettier --single-quote --check src",
     "prettier": "prettier --single-quote --write .",
-    "gh-pages": "react-scripts build && gh-pages -d build"
+    "gh-pages": "tsc && vite build && gh-pages -d build"
   },
   "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest",
-      "prettier"
-    ]
+    "extends": []
   },
   "browserslist": {
     "production": [
@@ -71,32 +90,63 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/plugin-transform-runtime": "^7.24.7",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-commonjs": "^25.0.8",
+    "@tapis/tapisui-extension-devtools": "file:./packages/tapisui-extension-devtools",
+    "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/react-hooks": "^7.0.2",
     "@testing-library/user-event": "^12.8.3",
+    "@types/base-64": "^1.0.0",
     "@types/jest": "^26.0.24",
     "@types/js-cookie": "^2.2.7",
     "@types/lodash.clonedeep": "^4.5.6",
-    "@types/node": "^12.20.37",
-    "@types/react": "^17.0.17",
+    "@types/node": "^18.19.44",
+    "@types/normalize-path": "^3.0.2",
+    "@types/react": "^18.3.3",
     "@types/react-css-modules": "^4.6.3",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-redux": "^7.1.16",
     "@types/react-router-dom": "^5.1.8",
     "@types/react-table": "^7.7.7",
-    "@types/redux-logger": "^3.0.8",
-    "@types/redux-mock-store": "^1.0.3",
-    "@types/redux-saga": "^0.10.5",
-    "@types/redux-thunk": "^2.1.0",
     "@types/rx": "^4.1.2",
     "@types/streamsaver": "^2.0.1",
     "@types/uuid": "^8.3.1",
+    "chokidar": "^3.6.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "^8.3.0",
-    "gh-pages": "^3.2.3",
+    "gh-pages": "^6.1.1",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "path": "^0.12.7",
     "prettier": "^2.3.2",
-    "redux-mock-store": "^1.5.4",
-    "redux-saga-test-plan": "^4.0.3"
+    "rollup-plugin-visualizer": "^5.12.0",
+    "ts-jest": "^29.1.4",
+    "typescript": "^5.4.5",
+    "vite-plugin-commonjs": "^0.10.1",
+    "vite-plugin-env-compatible": "^2.0.1",
+    "vite-plugin-svgr": "^4.2.0",
+    "vite-tsconfig-paths": "^4.3.2"
+  },
+  "optionalDependencies": {
+    "@rollup/rollup-linux-x64-gnu": "4.9.5"
+  },
+  "jest": {
+    "globals": {
+      "Uint8Array": "Uint8Array"
+    },
+    "preset": "ts-jest/presets/js-with-ts",
+    "testEnvironment": "jest-environment-jsdom",
+    "testPathIgnorePatterns": [
+      "lib"
+    ],
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      ".+\\.(css|styl|less|sass|scss)$": "identity-obj-proxy"
+    },
+    "modulePaths": [
+      "<rootDir>/src"
+    ]
   }
 }
