@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Router } from 'react-router-dom';
 import { render } from '@testing-library/react';
-import TapisProvider from 'tapis-hooks/provider';
+import { TapisProvider } from '@tapis/tapisui-hooks';
 export default function renderComponent(component: any, history: any = null) {
   if (history) {
     return render(
